Release/0.6.0 (#51)
* Services and Mappers

* - Added Swagger.
- Fixed endpoints to conform to standards.

* XML documentation for swagger

* New service interfaces. Paging!

* Animal paging back-end

* FE paging

* Paging and reworking styling

* Reworking theming

* Reworked layout to use mat-sidenav. Also added a card around the
dashboard info.

* Animal list: now a table

* Animal List: row display

* Styling: container

* Animals Controller: Documented

* Animal: Crud service

* Animal CRUD service

* SQLite middleware extensions and documentation

Paging for feed types. Database standardisation.

Paging: Feed Type front-end

Feeding Transaction: FE rework

Fixed feeding transactions not being fetched

* Paging: Feeding transactions

* Fixed feeding transaction CRUD issues

* Unit service

* Paging: Units back-end

* Pagination: units

* CRUD services and ngrx store: Medicine types

* Added XML documentation to MedicineTypeController

* Pagination and CRUD services: Medical transaction

* Remove dangling service files

* Fixed medical transaction update API endpoint

* ngrx store: Medical transactions

* Front End: Medical transaction list rework

* Medical module: lazy loading

* Medical module: medical transaction form

* Bump websocket-extensions in /livestock-tracker/ClientApp (#24)

Bumps [websocket-extensions](https://github.com/faye/websocket-extensions-node) from 0.1.3 to 0.1.4.
- [Release notes](https://github.com/faye/websocket-extensions-node/releases)
- [Changelog](https://github.com/faye/websocket-extensions-node/blob/master/CHANGELOG.md)
- [Commits](https://github.com/faye/websocket-extensions-node/compare/0.1.3...0.1.4)



* I can't remember

* Feeding transaction conforms to new pattern

* Reworked animal form screen. Fixing issue with update animals where only summary data is in memory

* Bump acorn from 6.4.0 to 6.4.2 in /livestock-tracker/ClientApp (#31)

Bumps [acorn](https://github.com/acornjs/acorn) from 6.4.0 to 6.4.2.
- [Release notes](https://github.com/acornjs/acorn/releases)
- [Commits](https://github.com/acornjs/acorn/compare/6.4.0...6.4.2)



* Bump http-proxy from 1.18.0 to 1.18.1 in /livestock-tracker/ClientApp (#29)

Bumps [http-proxy](https://github.com/http-party/node-http-proxy) from 1.18.0 to 1.18.1.
- [Release notes](https://github.com/http-party/node-http-proxy/releases)
- [Changelog](https://github.com/http-party/node-http-proxy/blob/master/CHANGELOG.md)
- [Commits](https://github.com/http-party/node-http-proxy/compare/1.18.0...1.18.1)



* Bump elliptic from 6.5.1 to 6.5.3 in /livestock-tracker/ClientApp (#27)

Bumps [elliptic](https://github.com/indutny/elliptic) from 6.5.1 to 6.5.3.
- [Release notes](https://github.com/indutny/elliptic/releases)
- [Commits](https://github.com/indutny/elliptic/compare/v6.5.1...v6.5.3)



* Bump lodash from 4.17.15 to 4.17.19 in /livestock-tracker/ClientApp (#26)

Bumps [lodash](https://github.com/lodash/lodash) from 4.17.15 to 4.17.19.
- [Release notes](https://github.com/lodash/lodash/releases)
- [Commits](https://github.com/lodash/lodash/compare/4.17.15...4.17.19)



* Bump npm-registry-fetch in /livestock-tracker/ClientApp (#25)

Bumps [npm-registry-fetch](https://github.com/npm/registry-fetch) from 4.0.2 to 4.0.5.
- [Release notes](https://github.com/npm/registry-fetch/releases)
- [Changelog](https://github.com/npm/npm-registry-fetch/blob/latest/CHANGELOG.md)
- [Commits](https://github.com/npm/registry-fetch/commits)



* Fix npm package vulnerabilities

* Upgrade packages to latest stable of same major version

* Add storybook (#34)

- Update npm packages
- Upgrade Angular to v9i
- Add Storybook

* Feature/loader (#35)

* Loader

* Fixed animal loading

* Fixed loader for animal list (#36)

* feature/animal-v2 (#37)

* feature/angular-v10 (#38)

* feature/angular-v11 (#39)

* ESLint: fixed errors and warnings (#40)

Changed a lot of files to be consistent with styling

* Feature/dotnet 5 (#42)

* .NET 5

* Fix tests

* Added one more unit test

* Updater converted to .NET 5

* Weight transactions (#48)

* Weight transaction models

* Fixed date migrations

* Weight services

* API endpoints for Weight Transactions

* Weight FE scaffolding

* Weigh transaction unit tests

* Weight state and unit test fixes

* Weight transaction form

The weight transaction form is a component that is focussed
on capturing the information relating to a weight transaction.
It will be hosted on both new and edit pages. It is a pure
component only concerned with rendering logic.

The pages, edit and new are located at routes:
- new: `/weight/new`
- edit: `/weight/:id/edit`
They are responsible for managing the data fetch requests
and async operations.

* Angular v12 (#49)

* Updated Storybook to v6.2.9

* Upgraded Angular core modules to v12

* Upgrade Angular Material to v12

* Upgraded NgRx to v12

* Update packages

* Update dotnet packages to latest minor (#50)

+++ livestock-tracker.abstractions/Models/Feed/IFeedingTransaction.cs
@@ -10,12 +10,12 @@ namespace LivestockTracker.Abstractions.Models.Feed
         /// <summary>
         /// The number that uniquely identifies the feeding transaction.
         /// </summary>
-        int Id { get; }
+        long Id { get; }
 
         /// <summary>
         /// A linking ID to the animal that was fed.
         /// </summary>
-        int AnimalId { get; }
+        long AnimalId { get; }
 
         /// <summary>
         /// A linking ID to the type of feed.
