+++ livestock-tracker.logic/Services/Medical/MedicalTransactionCrudService.cs
@@ -4,6 +4,7 @@ using LivestockTracker.Abstractions.Models.Medical;
 using LivestockTracker.Abstractions.Services.Medical;
 using LivestockTracker.Database;
 using LivestockTracker.Database.Models.Medical;
+using LivestockTracker.Logic.Paging;
 using LivestockTracker.Models.Medical;
 using LivestockTracker.Models.Paging;
 using Microsoft.EntityFrameworkCore;
@@ -170,7 +171,7 @@ namespace LivestockTracker.Logic.Services.Medical
         /// <param name="key">The unique identifier value for the medical transaction.</param>
         /// <param name="cancellationToken">A token that can be used to signal operation cancellation.</param>
         /// <returns>The matching medical transaction if found.</returns>
-        public virtual async Task<IMedicalTransaction?> GetOneAsync(int key, CancellationToken cancellationToken)
+        public virtual async Task<IMedicalTransaction?> GetOneAsync(long key, CancellationToken cancellationToken)
         {
             Logger.LogInformation($"Finding a medical transaction that matches ID {key}...");
             var medicalTransaction = await LivestockContext.MedicalTransactions
@@ -187,7 +188,7 @@ namespace LivestockTracker.Logic.Services.Medical
         /// <param name="key">The unique identifier value for the medical transaction.</param>
         /// <param name="cancellationToken">A token that can be used to signal operation cancellation.</param>
         /// <returns>The ID of the removed medical transaction.</returns>
-        public virtual async Task<int> RemoveAsync(int key, CancellationToken cancellationToken)
+        public virtual async Task<long> RemoveAsync(long key, CancellationToken cancellationToken)
         {
             Logger.LogInformation($"Deleting medical transaction with ID {key}...");
             var medicalTransaction = await LivestockContext.MedicalTransactions
